Part 1 - Academic Progression Marker (Main)
This section mostly concentrated on guiding the user to the specified credit score (pass, defer, or fail). It also checked the user-input data type and entered marks to ensure that they were within the range. If the credit score exceeded the total marks, an error notice had to be displayed. and this portion concentrated on various outcomes as well (checking an infinite number of students) after examining one student's progression.